Revert "moved shared content key"
This reverts commit a9156a61fb7ade904d7f19f8cbfc1d97dcb65421.
